+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 -   Added ability to set all settings to defaults.
 -   Added google analytics to measure if people use it or not. Feel free to block this script, and let me know if you want it to be removed.
 -   Tweaked weather CSS.
+-   Added the ability to customise search suggestion colours.